Added method to get data from google sheet

+++ src/main/java/aupet/microclimate/model/enums/StandartCode.java
@@ -5,13 +5,13 @@ public enum StandartCode {
     TEMPERATURE_STAND(19),
 
     ;
-    private final int code;
+    private final double code;
 
-    StandartCode(int code) {
+    StandartCode(double code) {
         this.code = code;
     }
 
-    public int getCode() {
+    public double getCode() {
         return code;
     }
 }
